% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_norm.R
\name{check_norm}
\alias{check_norm}
\title{Check Normality}
\usage{
check_norm(resid, grid = 3)
}
\arguments{
\item{resid}{Residuals from a fit model}

\item{grid}{Dimension for the grid of plots, given as an integer.}
}
\description{
This function is designed to allow users to check the normality of residuals through an intuitive approach. It generates a random set of normally distributed variables, then inserts the given residuals into this set. If users can distinguish the QQ Plot of their data from the randomly generated normal data, their data does not follow a normal distribution.
}
\details{

}
\author{
Michael Floren
}
