% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ept.R
\name{ept}
\alias{ept}
\title{Eval Parse Text}
\usage{
ept(x)
}
\arguments{
\item{x}{This is the text to parse}
}
\description{
This is a short function which implements eval(parse(text=__)) quickly (including a minor error check (that the blank is text)).
}
\details{
Not much to it! Currently requires a character.
}
\examples{
#this gives 1
a=1
ept("a")

}
\author{
Michael Floren
}
