% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_ind_p.R
\name{fix_ind_p}
\alias{fix_ind_p}
\title{Fix Individual P-Values}
\usage{
fix_ind_p(x, dig = 2, wp = TRUE, equals = TRUE)
}
\arguments{
\item{x}{The p-value to "fix"}

\item{dig}{The number of digits you wish to round to}

\item{wp}{Should the letter "p" be included?}
}
\description{
A function to fix individual p-values for reporting
}
\details{
This function takes an individual pvalue and formats it appropriately. The letter p can be included or not, depending on what is desired for the report.
}
\examples{
fix_ind_p(0.05, dig=2)
fix_ind_p(0.005, dig=2)
fix_ind_p(0.05, dig=2, wp=FALSE)

}
\seealso{
\code{\link[=fixp]{fixp()}}
}
\author{
Michael Floren
}
