% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr.R
\name{fr}
\alias{fr}
\title{Format and Round}
\usage{
fr(x, dig = 2, remlead0 = FALSE, trim = TRUE, ...)
}
\arguments{
\item{x}{The value to round and format}

\item{dig}{The number of digits you wish to round to}

\item{...}{Additional arguments passed to \code{format}}
}
\description{
A function to round and format decimals to the correct digits (including following zeros).
}
\details{
This function expects a number. It uses the round command to round appropriately, then the format command to print a string with the correct number of digits. The output is a string.
}
\examples{
fr(2.04, 2)
fr(2.04, 1)
fr(2.04, 0)

}
\author{
Michael Floren
}
