% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab.R
\name{grab}
\alias{grab}
\title{The Grab Function}
\usage{
grab(..., list = character(), dependencies = TRUE)
}
\arguments{
\item{...}{Packages that should be loaded}

\item{list}{A character list of packages that should be loaded}

\item{dependencies}{Passed to \code{install.packages}. Default is \code{TRUE}}
}
\description{
This function is designed to be used in place of library and install.packages. It will check if a potential package is downloaded and, if not, it will try to download it using install.packages.
}
\details{
The function uses the logical output from \code{require} to determine if the package exists locally. If it does, the package is then loaded using \code{library}. If not, the function tries do download the package using \code{install.packages}, checks to make sure it downloaded, then loads the package.
}
\seealso{
\code{\link[base]{install.packages}, \code{\link[base]{require}}, \code{\link[base]{library}}}
}
\author{
Michael Floren
}
