% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midval.R
\name{midval}
\alias{midval}
\title{Middle Value}
\usage{
midval(x)
}
\arguments{
\item{x}{Vector of interest}
}
\description{
Find the middle value of a vector
}
\details{
Simply finds the middle value, using the first one if there are 2 (i.e., for an even lengthed vector).
}
\examples{
midval(1:10)
midval(1:11)

}
\author{
Michael Floren
}
