% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/px.R
\name{px}
\alias{px}
\title{Print X-Table}
\usage{
px(x, cap = NULL, lab = NULL, rn = FALSE, cn = TRUE, scale = "1", ...)
}
\arguments{
\item{x}{The table to be written to LaTeX}

\item{cap}{The caption of the table}

\item{lab}{The label to be assigned}

\item{rn}{Logical. Do you wish to print row names?}

\item{cn}{Logical. Do you wish to print column names?}

\item{scale}{A scaling factor (smaller scale = smaller text)}

\item{...}{Additional arguments (passed to xtable and print.xtable)}
}
\description{
This function is intended to be a wrapper for printing xtables with some common options that I normally use and change. This provides no unique or creative commands, and is simply used for convenience.
}
\details{
px Version 0.1

Again, just combining print.xtable into one short function with some convenient defaults for my use.
}
\seealso{
\code{\link[xtable]{print.xtable}}
}
\author{
Michael Floren
}
