% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txtme.R
\name{txtme}
\alias{txtme}
\title{Text Me}
\usage{
txtme(txt = "", sbj = "", resetfrom = FALSE, resetto = FALSE,
  resetpw = FALSE, resetall = FALSE, def.car = "@tmomail.net",
  keeptime = NULL)
}
\arguments{
\item{txt}{The message to text}

\item{sbj}{The subject of the message}

\item{resetfrom}{Logical. If true, resets "from"}

\item{resetto}{Logical. If true, resets "to"}

\item{resetpw}{Logical. If true, resets the password}

\item{resetall}{Logical. If true, reset "to", "from", and password}

\item{keeptime}{Numeric (with character option). Specify how long you would like your password to be stored (in minutes). Note that, when reset, your password will need to be reset. Default is 10 minutes. Entry of text string "forever" will keep password with no time-limit.}

\item{defcar}{This the ending for the carrier you want to text}
}
\description{
This function is designed to send a user a text message (currently, settings are designed to do this from a G-Mail account) to a designated phone number (current carrier is set to T-Mobile). It will store the password for a set amount of time (defaulting to 10 minutes) before requiring it to be re-entered
}
\details{
This fuunction uses the memoise package to store passwords as variables in the global environment. This may or may not be desireable, but was deemed most appropriate of potential options. Please see the memoise documentation for details on this storage (read: due to my ignorance, I really don't know). Note that, due to trial and error, passwords should be saved even if console is closed (likely saved in workspace). Again, this may be undesireable if \code{keeptime} is set to "forever", so it is recommended to build in password resets into chunks using this function if using code{keeptime="forever"}. 

The texting is truly done through sending an e-mail to a cell-phone's e-mail address: number@carrierinfo.blah... For those looking to change the carrier, please visit the following website to find your carrier's information: http://www.emailtextmessages.com/
}
\author{
Michael Floren
}
\seealso{
\code{\link[memoise]{memoise}}, \code{mailR}
}

