% Generated by roxygen2 (4.0.2): do not edit by hand
\name{keep}
\alias{keep}
\title{The Keep Function}
\usage{
keep(x = c())
}
\arguments{
\item{x}{A character vector of variables which you wish to keep}
}
\description{
The Keep Function
}
\details{
This function is designed to keep a given list of variables (as opposed to the rm function, which is designed to remove a list of variables)
}

