% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep.R
\name{keep}
\alias{keep}
\title{The Keep Function}
\usage{
keep(..., x = c())
}
\arguments{
\item{...}{Variables that should stay in the environment}

\item{x}{A character vector of variables which you wish to keep}
}
\description{
This function is designed to keep a given list of variables (as opposed to the rm function, which is designed to remove a list of variables)
}
\details{
The function uses the \code{ls} command to find the variables that are defined in the parent environment. It then deletes all variables that are not given as an argument.
}
\seealso{
\code{\link[base]{rm}}
}
\author{
Michael Floren
}
