% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/myStuff.R
\docType{package}
\name{myStuff}
\alias{myStuff}
\alias{myStuff-package}
\title{My Stuff}
\arguments{
\item{set_param}{This function can be used to set global statistical parameters from all in one place. When given N, n, and alpha, it will set those as global parameters as well as calculating z-scores and t-scores for the given population.}

\item{keep}{The keep function was designed to work in much the same was as rm, save to keep objects as opposed to removing them.}
}
\description{
This is a package that contains some useful functions that were originally created for me and my personal needs.
}
\details{


As a side note, the \code{keep} function was designed out of several projects I had where I moved quickly between them. I wanted to clear the workspace to avoid naming confusion, but I also wanted several variables to be left alone. \code{keep} can remove all variables from the workspace other than the ones listed
}

